not running hashbang test if the command would hit kernel length limit

+++ tests/hashbang.sh
@@ -7,6 +7,9 @@ echo '(lambda (args) (print (foldr string-append "" (list "#!" (foldr (lambda (a
 
 chmod +x tmp/script-$ME
 
+# hashbangs have a fairly short maximum length set by kernel
+test $(head -n 1 ./tmp/script-$ME | wc -c) -gt 127 && { echo ohai; exit 0; }
+
 ./tmp/script-$ME | grep "^ohai$" || exit 1
 
 rm tmp/script-$ME
